refactor(workspace): split shell search sync
