+++ libs/workspace/shell/feature/src/lib/workspace-shell/services/workspace-shell.facade.ts
@@ -213,10 +213,6 @@ export class WorkspaceShellFacade {
         this.header.refreshCurrentPlaylist();
     }
 
-    private syncSearchFromRoute(): void {
-        this.search.syncSearchFromRoute();
-    }
-
     private makeCommandBuilderContext(): CommandBuilderContext {
         return {
             route: this.currentRoute(),
